Update schema version info message when importing a new RT/profile
Fixes #396

This introduces a new env-overrideable config variable setting the default RT/Profile schema version. Note that I was unable to determine whether 0.0.9 is the right default for now. #396 specifies 0.0.9 so that is what I went with, but 0.1.0 seems to be the newest version available. FWIW, I attempted to upload a few seemingly valid 0.1.0 and 0.0.9 profiles and I was unable to get *any* to validate. I am not sure whether that should be in scope for this issue but I would at least like to understand why this is the case.