Bump sinopia_server to 3.0.0-beta4
Fixes #506

Using the newer version of the sinopia client allows the editor to invoke the updateResource API with a custom content type (application/json). Prior to this change, updates worked (apparently) but what was happening was the editor was issuing PUT requests and getting 204s... but not modifying the NonRDFSource. Instead the RDFSource that Trellis auto-attaches to every NonRDFSource was being modified.