Check whether server has new templates when updating SinopiaResourceTemplates
Fixes #524

This PR adds a new condition to `SinopiaResourceTemplates.componentDidUpdate()` that issues a HEAD request to the LD4P group to get its ETag, and compares it against the ETag stored in component state. If the ETag has changed, it means the LD4P group has changed which in practice is a signal that one or more resource templates have been added to the group since the component loaded. This makes sure that server conditions can defeat local caching that is currently preventing sessions from seeing templates created in other sessions (without using a hard refresh).