Extract the mapping functions from state to properties so they can be tested
One option for testing mapStateToProps directly would have required exporting private code as public, just for the purpose of testing.  That's a bit of a code smell.

Another option would be to pass a mock store into the component.  That was something we could do in react-redux v5 and which was restored in v7, but it was removed in v6, which is what we are on. See https://github.com/reduxjs/react-redux/issues/1161

This seems like the best way of testing this mapping without upgrading react-redux or unnecessarily breaking encapsulation.