Use webpack to build both dev and prod servers
And use webpack-dev-middleware instead of webpack-dev-server, so that the dev server can make use of the same Express server used in prod. This brings dev and prod configurations closer together.

Refs #630

This is the first step in tackling #630, exposing a ElasticSearch proxy via the application express server. This work aligns our production server infrastructure with dev server infrastructure, making sure that when the proxy is built, it is exposed identically in both dev and prod.