Extract a functional component for rendering search result rows
This is a first pass at extracting the functional component. Would like to get an early review of this approach to make sure I am on the right track before I add/change tests for the React components. I am particularly interested in hearing whether I should be using props and/or state differently in the two involved components.