Persist RDF to Trellis using Turtle
Fixes #1753

Big thanks to @justinlittman for sharing the writable stream technique; that was
the key to convincing the N3Writer to operate in synchronous mode (which wasn't
obvious to me from the N3 documentation, which is otherwise quite good). To be
discussed/tested before merging and going with this approach:

1. With the changes in this branch, how does/will the editor handle the currently
broken triples coming back from Trellis? (I *think* they will load since we have
code in place for handling null subjects in N-Triples data, but I could be
mistaken.
2. Assuming all is copacetic with (1) above, will the code in this branch
automagically handle migration of data that needs fixing, or do we need a more
intentional and systematic migration strategy for all the broken data?
3. When this is deployed, what effect will it have on open
sessions (particularly w/ regard to the changes in this branch related to
checking to see if a resource has changed, which used to do this check by
hashing the N-Triples representation but now does it by hashing the Turtle
representation)?