Auto-resize input literals to reduce scrolling
Fixes #2177

Note that a UI change here is switching from an input element to a textarea element for all input literals. Why? Because as far as I can tell, input elements are not auto-resizable in HTML.

I briefly looked into rolling my own auto-resize logic but it is over my head, even without all the React machinery in the way. I also looked into using contentEditable but I believe that approach is geared towards divs and spans rather than input elements (such as textarea) within a form.

I'm not sure if the small test change is sufficient to test this commit. Advice welcome! I suspect that the line I removed from the input literal test, which began failing, is due to moving from an input element to a textarea element.