Make resource close button close the expected tab
Fixes #2057

The prior implementation of CloseResourceModal used the selectCurrentResourceKey to figure out which resource the CloseButton should close, and that was the bug. Instead, the CloseButton should tell the CloseResourceModal what its resource is and it should close that resource.

As for testing, this change was tested in the browser and confirmed working. See linked screen capture:

...

I briefly entertained the idea of adding unit tests for CloseButton and CloseResourceModal but it seemed of little value to test properties injection which is core React functionality. Given also that 1) such tests would not make it apparent that this bug is related to the tests, 2) my familiarity with React testing is relatively low, and 3) this is fixing a production bug, I decided not to add any automated tests in this commit. Happy to reconsider.