Merge groups into resource only when not a new resource
Fixes #3199

This commit makes a small tweak to the spot in the codebase where groups are merged into resources, short-circuiting that behavior when the resource is new. The intent is to prevent copied resources from inheriting groups and edit groups from the resources from which they were copied. These should not be set, and the user should select the group and edit groups after copying the resource.