add authority humord and fix ntsf
Humord is a request for an authority by the Norwegian Team.  Issue (https://github.com/LD4P/qa_server/issues/483)

The changes for NTSF reflect a copy/paste error made in the QA config.  This authority does not in actuality have any sub-authorities. 

Both these changes are in production QA.